+++ lib/Castle.php
@@ -18,6 +18,7 @@ if (!function_exists('lcfirst'))
 require(dirname(__FILE__) . '/Castle/Castle.php');
 require(dirname(__FILE__) . '/Castle/CookieStore.php');
 require(dirname(__FILE__) . '/Castle/Errors.php');
+require(dirname(__FILE__) . '/Castle/Failover.php');
 require(dirname(__FILE__) . '/RestModel/Resource.php');
 require(dirname(__FILE__) . '/RestModel/Model.php');
 require(dirname(__FILE__) . '/Castle/Context.php');